requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.9.24
python-telegram-bot==13.14
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tornado==6.1
tzdata==2022.5
tzlocal==4.2
psycopg2-binary==2.9.3
pycodestyle==2.8.0
python-dotenv==0.20.0
requests==2.28.1
SQLAlchemy==1.4.36

#CREATE EXTENSION pg_trgm;